% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_time.R
\name{filter_by_time}
\alias{filter_by_time}
\title{Filter Recipes by Preparation and Cooking Time}
\usage{
filter_by_time(
  data,
  min_prep_time = 0,
  max_prep_time = Inf,
  min_cook_time = 0,
  max_cook_time = Inf
)
}
\arguments{
\item{data}{Data frame containing the recipes.}

\item{min_prep_time}{Minimum allowed preparation time in minutes (default is 0).}

\item{max_prep_time}{Maximum allowed preparation time in minutes (default is Inf).}

\item{min_cook_time}{Minimum allowed cooking time in minutes (default is 0).}

\item{max_cook_time}{Maximum allowed cooking time in minutes (default is Inf).}
}
\value{
A data frame of recipes filtered by the specified time constraints.
}
\description{
This function filters recipes based on specified minimum and maximum preparation and cooking times.
}
\examples{
filter_by_time(data = recipes, min_prep_time = 10, max_prep_time = 30,
               min_cook_time = 20, max_cook_time = 60)
}
