% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_recipes}
\alias{get_recipes}
\title{Filter Recipes Based on Various Criteria}
\usage{
get_recipes(
  data,
  ingredients = NULL,
  cuisine = NULL,
  diet = NULL,
  prep_time_max = Inf,
  cook_time_max = Inf
)
}
\arguments{
\item{data}{Data frame containing the recipes.}

\item{ingredients}{Vector of ingredients for filtering (optional).}

\item{cuisine}{String specifying the cuisine type for filtering (optional).}

\item{diet}{String specifying the diet type for filtering (optional).}

\item{prep_time_max}{Numeric value for maximum preparation time in minutes (optional).}

\item{cook_time_max}{Numeric value for maximum cooking time in minutes (optional).}
}
\value{
A data frame of filtered recipes.
}
\description{
This function filters the recipes dataset based on provided criteria such as ingredients,
cuisine, diet, maximum preparation time, and maximum cooking time.
}
\examples{
get_recipes(data, ingredients = c("tomato", "chicken"),
           cuisine = "Italian", diet = "Gluten Free",
           prep_time_max = 30, cook_time_max = 60)
}
