% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_n_recipes.R
\name{get_top_n_recipes}
\alias{get_top_n_recipes}
\title{Get Top N Rated Recipes}
\usage{
get_top_n_recipes(data, n = 5)
}
\arguments{
\item{data}{Data frame containing the recipes.}

\item{n}{The number of top-rated recipes to return.}
}
\value{
A data frame containing the top N rated recipes.
}
\description{
This function retrieves the top N rated recipes from the dataset based on user ratings.
}
\examples{
get_top_n_recipes(data = recipes, n = 5)
}
