% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfuction.R
\name{launchApp}
\alias{launchApp}
\title{Launch the Meal Suggestor App}
\usage{
launchApp()
}
\description{
This function starts the Meal Suggestor Shiny application, providing users
with a user-friendly interface to get personalized meal suggestions based
on their preferences and dietary requirements.
}
\examples{
To launch the Meal Suggestor application:
}
