% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaunchMealApp.R
\name{launchMealApp}
\alias{launchMealApp}
\title{Launch Meal Suggestor Shiny App}
\usage{
launchMealApp()
}
\description{
This function starts the Meal Suggestor Shiny application contained within the package.
It assumes that the Shiny app's UI and server logic are defined in an external script named 'app.R'.
}
