% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_recipe.R
\name{random_recipe}
\alias{random_recipe}
\title{Select a Random Recipe}
\usage{
random_recipe(data)
}
\arguments{
\item{data}{Data frame containing the recipes.}
}
\value{
A single randomly selected recipe.
}
\description{
This function selects a random recipe from the datase to suprise the customer
}
\examples{
random_recipe(data = recipes)
}
