% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{get_recipes}
\alias{get_recipes}
\title{Filter Recipes Based on Various Criteria}
\usage{
get_recipes(data, ingredients = NULL, cuisine = NULL, diet = NULL)
}
\arguments{
\item{data}{Data frame containing the recipes.}

\item{ingredients}{Vector of ingredients for filtering}

\item{cuisine}{String specifying the cuisine type for filtering}

\item{diet}{String specifying the diet type for filtering}
}
\value{
A data frame of filtered recipes.
}
\description{
This function filters the recipes dataset based on provided criteria such as ingredients,
cuisine, diet, maximum preparation time, and maximum cooking time.
}
\examples{
get_recipes(data, ingredients = c("tomato", "chicken"),
           cuisine = "Italian", diet = "Gluten Free")
}
